requirements.txt
streamlit
langchain==0.3.13
langchain_community
langchain_core
python-dotenv
langchain-huggingface
langchain-qdrant
langchain-ollama
unstructured[pdf]
onnx==1.16.1
qdrant-client
openpyxl
python-docx
pandas
tqdm